@@ -1,10 +1,19 @@
-use crate::{ServerJSONResp, Str};
+use crate::ServerJSONResp;
 use serde::{Deserialize, Serialize};
-use serde_json::from_str;
+use serde_json::{from_str, to_string};
 use std::{collections::HashMap, env::consts::ARCH};
 
 pub mod install;
 mod other_fields;
+
+#[cfg(feature = "js")]
+use tsify::*;
+
+#[cfg(feature = "js")]
+use wasm_bindgen::JsValue;
+
+#[cfg(feature = "js")]
+use js_macros::TsifyAsync;
 
 pub use install::*;
 pub use other_fields::*;
@@ -12,18 +21,23 @@ pub use other_fields::*;
 #[allow(non_snake_case)]
 #[derive(Serialize, Deserialize, Debug, Clone)]
 #[doc = "Use the official ahqstore (https://crates.io/crates/ahqstore_cli_rs) cli\n🎯 Introduced in v1, Revamped in v2"]
+#[cfg_attr(feature = "js", derive(Tsify, TsifyAsync))]
+#[cfg_attr(feature = "js", tsify(into_wasm_abi, from_wasm_abi))]
 pub struct AHQStoreApplication {
   /// The ID of the application
-  pub appId: Str,
+  pub appId: String,
 
   /// The name of the shortcut of the app
-  pub appShortcutName: Str,
+  pub appShortcutName: String,
 
   /// The name that'll be displayed in the app
-  pub appDisplayName: Str,
+  pub appDisplayName: String,
 
   /// Unique ID of the author
-  pub authorId: Str,
+  pub authorId: String,
+
+  /// The TagName of the release, MUST NOT BE LATEST
+  pub releaseTagName: String,
 
   /// Download URLs that the app will address
   pub downloadUrls: HashMap<u8, DownloadUrl>,
@@ -31,37 +45,114 @@ pub struct AHQStoreApplication {
   /// Install options
   pub install: InstallerOptions,
 
-  /// App display images, let the cli do it
-  pub displayImages: Vec<Str>,
+  /// App display images referencing /resources, let the cli do it
+  pub displayImages: Vec<u8>,
 
   /// App description
-  pub description: Str,
-
-  /// App Icon (base64), let the cli do it
-  pub icon: Str,
+  pub description: String,
 
   /// Your Github Repo associated
   pub repo: AppRepo,
 
   /// Will be generated by the cli, it is the unique-version-hash of your app
-  pub version: Str,
+  pub version: String,
 
   /// The Site to your app
-  pub site: Option<Str>,
+  pub site: Option<String>,
 
   /// This'll be ignored unless you're ahq_verified tag which no one except AHQ Store Team has
-  /// 
+  ///
   /// The general dev isn't meant to redistribute others' apps unless they own right to do so
-  pub source: Option<Str>,
+  pub source: Option<String>,
 
   /// License type or Terms of Service Page
-  pub license_or_tos: Option<Str>,
+  pub license_or_tos: Option<String>,
 
   /// Page of Application
-  pub app_page: Option<Str>,
+  pub app_page: Option<String>,
+
+  /// These Resources will be passed to the installer, the size of all the Vec<u8> must not be more than 5 * 1024 * 1024 * 1024 bytes (~5MB)
+  pub resources: Option<HashMap<u8, Vec<u8>>>,
 }
 
 impl AHQStoreApplication {
+  #[cfg(feature = "apps_repo")]
+  pub fn validate(&self) -> Result<String, String> {
+    let mut result = String::new();
+
+    result.push_str(&self.validate_resource());
+
+    if &self.authorId != "1" {
+      if &self.releaseTagName == "latest" {
+        result.push_str("❌ ReleaseTagName can't be latest\n");
+      }
+
+      if let Some(_) = self.source {
+        result.push_str("❌ Source can't be present, your application must be original\n");
+      }
+    }
+
+    if result.contains("❌") {
+      Err(result)
+    } else {
+      Ok(result)
+    }
+  }
+
+  #[cfg(feature = "apps_repo")]
+  pub fn validate_resource(&self) -> String {
+    let Some(x) = &self.resources else {
+      return "❌ No Resources Present".into();
+    };
+
+    let mut total = 0;
+
+    x.iter().for_each(|(_, v)| total += v.len());
+
+    let mut resp = String::new();
+
+    if total > 5 * 1024 * 1024 * 1024 {
+      resp.push_str("❌ Total size of all resources combined must not be more than 5MB\n");
+    } else if self.displayImages.len() > 6 {
+      resp.push_str(
+        "❌ A maximum of 6 images (1 icon + 5 display images) can be set in displayImages\n",
+      );
+    } else if x.get(&0).is_none() {
+      resp.push_str("❌ Resource with id 0 must be present as it represents icon\n");
+    } else if !self
+      .displayImages
+      .iter()
+      .all(|id| x.get(&(*id + 1)).is_some())
+    {
+      resp.push_str("❌ Every display images should have their resource id\n");
+    } else {
+      resp.push_str("✅ Resources are valid\n");
+    }
+
+    resp
+  }
+
+  pub fn export(&self) -> (String, Vec<(u8, Vec<u8>)>) {
+    let mut obj = self.clone();
+
+    for val in obj.downloadUrls.values_mut() {
+      if &obj.authorId == "1" && &val.asset == "url" {
+        continue;
+      }
+
+      let path = format!(
+        "https://github.com/{}/{}/releases/download/{}/{}",
+        self.repo.author, self.repo.repo, self.releaseTagName, val.asset
+      );
+
+      val.url = path;
+    }
+    let resources: Vec<(u8, Vec<u8>)> = std::mem::replace(&mut obj.resources, None)
+      .map_or(vec![], |map| map.into_iter().collect::<Vec<_>>());
+
+    (to_string(&obj).unwrap(), resources)
+  }
+
   pub fn list_os_arch(&self) -> Vec<&'static str> {
     self.install.list_os_arch()
   }
@@ -74,8 +165,8 @@ impl AHQStoreApplication {
     self.install.has_platform()
   }
 
-  #[doc = "🎯 Introduced in v2"]
-  pub fn get_win_download(&self) -> Option<&DownloadUrl> {
+  #[doc = "🎯 Introduced in v3"]
+  pub fn get_win_options(&self) -> Option<&InstallerOptionsWindows> {
     let get_w32 = || {
       let Some(x) = &self.install.win32 else {
         return None;
@@ -95,6 +186,12 @@ impl AHQStoreApplication {
       get_w32()?
     };
 
+    Some(win32)
+  }
+
+  #[doc = "🎯 Introduced in v2"]
+  pub fn get_win_download(&self) -> Option<&DownloadUrl> {
+    let win32 = self.get_win_options()?;
     let url = self.downloadUrls.get(&win32.assetId)?;
 
     match &url.installerType {
@@ -129,16 +226,21 @@ impl AHQStoreApplication {
     self.get_win_extension()
   }
 
-  #[doc = "🎯 Introduced in v2"]
-  pub fn get_linux_download(&self) -> Option<&DownloadUrl> {
-    let linux = match ARCH {
-      "x86_64" => self.install.linux.as_ref()?,
-      "aarch64" => self.install.linuxArm64.as_ref()?,
-      "arm" => self.install.linuxArm7.as_ref()?,
+  #[doc = "🎯 Introduced in v3"]
+  pub fn get_linux_options(&self) -> Option<&InstallerOptionsLinux> {
+    match ARCH {
+      "x86_64" => self.install.linux.as_ref(),
+      "aarch64" => self.install.linuxArm64.as_ref(),
+      "arm" => self.install.linuxArm7.as_ref(),
       _ => {
         return None;
       }
-    };
+    }
+  }
+
+  #[doc = "🎯 Introduced in v2"]
+  pub fn get_linux_download(&self) -> Option<&DownloadUrl> {
+    let linux = self.get_linux_options()?;
 
     let url = self.downloadUrls.get(&linux.assetId)?;
 
