@@ -1,16 +1,23 @@
+#[cfg(feature = "js")]
+use wasm_bindgen::prelude::wasm_bindgen;
+
 use std::fmt::Display;
 
-use crate::Str;
 use serde::{Deserialize, Serialize};
 
 #[allow(non_snake_case)]
 #[derive(Serialize, Deserialize, Debug, Clone)]
+#[cfg_attr(feature = "js", wasm_bindgen(getter_with_clone))]
 pub struct DownloadUrl {
   pub installerType: InstallerFormat,
-  pub url: Str,
+  pub asset: String,
+
+  /// This will be based on asset and releaseId
+  pub url: String,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
+#[cfg_attr(feature = "js", wasm_bindgen)]
 pub enum InstallerFormat {
   #[doc = "🎯 Stable as of v1"]
   WindowsZip,
@@ -51,15 +58,16 @@ impl Display for InstallerFormat {
         InstallerFormat::WindowsInstallerMsi => "Windows Installer Msi",
         InstallerFormat::WindowsUWPMsix => "UWP Windows Msix Package",
         InstallerFormat::LinuxAppImage => "Linux App Image",
-        InstallerFormat::AndroidApkZip => "Universal Android Apk Zip Package"
+        InstallerFormat::AndroidApkZip => "Universal Android Apk Zip Package",
       }
     )
   }
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
+#[cfg_attr(feature = "js", wasm_bindgen(getter_with_clone))]
 pub struct AppRepo {
   /// author must be your GitHub username or username of an org where you're a "visible" member
-  pub author: Str,
-  pub repo: Str,
+  pub author: String,
+  pub repo: String,
 }
