@@ -2,6 +2,6 @@ use std::thread::JoinHandle;
 
 pub type Malicious = bool;
 
-pub fn scan_threaded<T>(_p: &T) -> JoinHandle<Malicious> {
-  std::thread::spawn(|| false)
+pub fn scan_threaded<T>(_p: &T) -> JoinHandle<Option<Malicious>> {
+  std::thread::spawn(|| Some(false))
 }
