@@ -220,7 +220,8 @@ pub fn main() {
       is_development,
       check_install_update,
       show_code,
-      rem_code
+      rem_code,
+      hash_username
     ])
     .menu(|handle| Menu::new(handle))
     .build(context)
@@ -297,6 +298,11 @@ pub fn main() {
     },
     _ => {}
   });
+}
+
+#[tauri::command(async)]
+fn hash_username(username: String) -> String {
+  ahqstore_gh_hash::compute(username.as_str())
 }
 
 #[tauri::command(async)]
