@@ -66,11 +66,14 @@ export default function Search(props: SearchProps) {
             }`}
           >
             <span className="block">
-              {searched ? <>0 Apps Found</> : <div className="flex">
-                <span className="dui-loading dui-loading-spinner" />
-                <span className="ml-2">Just a moment...</span>
-              </div>
-              }
+              {searched ? (
+                <>0 Apps Found</>
+              ) : (
+                <div className="flex">
+                  <span className="dui-loading dui-loading-spinner" />
+                  <span className="ml-2">Just a moment...</span>
+                </div>
+              )}
             </span>
           </div>
         ) : (
@@ -105,10 +108,14 @@ export default function Search(props: SearchProps) {
             }`}
           >
             <span className="block">
-              {searched ? "0 Apps Found" : <div className="flex w-[100%] justify-center">
-                <span className="dui-loading dui-loading-spinner dui-loading-lg" />
-                <span className="ml-2">Just a moment...</span>
-              </div>}
+              {searched ? (
+                "0 Apps Found"
+              ) : (
+                <div className="flex w-[100%] justify-center">
+                  <span className="dui-loading dui-loading-spinner dui-loading-lg" />
+                  <span className="ml-2">Just a moment...</span>
+                </div>
+              )}
             </span>
           </div>
         ) : (
