@@ -1,11 +1,11 @@
-import { useRef } from "react";
+import { useEffect, useRef, useState } from "react";
 
 //Icons
 import { MdModeEdit } from "react-icons/md";
 import { IoIosNotifications } from "react-icons/io";
 
 //API
-import { appData } from "../../resources/api/fetchApps";
+import { appData, getResource } from "../../resources/api/fetchApps";
 import Toast from "../../resources/api/toast";
 import { invoke } from "@tauri-apps/api/core";
 
@@ -23,12 +23,20 @@ export default function App({
   const updating = false;
   const data = useRef<HTMLDivElement>("" as any);
 
+  const [icon, setIcon] = useState<string>();
+
   async function handleClick() {
     toast("Opened in browser", "success", 1);
     invoke("open", {
-      url: "https://github.com/ahqstore/data",
+      url: "https://github.com/ahqstore/apps",
     });
   }
+
+  useEffect(() => {
+    (async () => {
+      setIcon(await getResource(appInfo.appId, "0"));
+    })();
+  }, [appInfo.appId]);
 
   return (
     <div
@@ -36,17 +44,24 @@ export default function App({
         dark ? "bg-gray-800 text-white" : "bg-gray-100 text-slate-800"
       } ${lastIndex ? "rounded-b-md" : ""} hover:shadow-xl pl-2 cursor-default`}
     >
-      <img
-        src={appInfo.icon}
-        alt={appInfo.appDisplayName}
-        className={`mr-2`}
-        draggable={false}
-        style={{
-          width: "60px",
-          marginTop: "auto",
-          marginBottom: "auto",
-        }}
-      ></img>
+      {icon ? (
+        <img
+          src={icon}
+          alt={appInfo.appDisplayName}
+          className={`mr-2`}
+          draggable={false}
+          style={{
+            width: "60px",
+            marginTop: "auto",
+            marginBottom: "auto",
+          }}
+        />
+      ) : (
+        <div
+          className={`dui-loading dui-loading-lg dui-loading-ring mt-5 mr-2 mb-[0.75rem] ${dark ? "text-white" : ""
+            }`}
+        />
+      )}
 
       <div className="flex flex-col my-auto text-start">
         <h1 className={`flex ${dark ? "text-blue-400" : "text-blue-700"}`}>
@@ -68,7 +83,7 @@ export default function App({
       {!updating ? (
         <div className="ml-auto mr-3 my-auto" ref={data}>
           <button
-            className="flex min-w-[100%] p-4 min-h-[3rem] justify-center items-center text-center text-blue-800 hover:text-white hover:bg-blue-800 rounded-xl transition-all cursor-pointer"
+            className="flex min-w-[100%] p-4 min-h-[3rem] justify-center items-center text-center text-base-content hover:text-accent-content hover:bg-accent-foreground rounded-xl transition-all cursor-pointer"
             onClick={() => {
               handleClick();
             }}
