@@ -26,7 +26,7 @@ export default function PopUp(props: PasswordProps) {
       minHeight: height || "40rem",
       transition: "all 500ms linear",
       backgroundColor: "var(--fallback-b1,oklch(var(--b1) / 1)",
-      border: "none"
+      border: "none",
     },
     overlay: {
       backgroundColor: "var(--fallback-b1,oklch(var(--b1) / 0.8)",
