@@ -3,6 +3,7 @@ import { ClientOptions, fetch as tauriFetch } from "@tauri-apps/plugin-http";
 export default async function fetch(
   url: string,
   config: (RequestInit & ClientOptions) | undefined,
+  mutate = true,
 ) {
   return await tauriFetch(url, {
     ...(config || {}),
@@ -13,14 +14,21 @@ export default async function fetch(
       "ngrok-skip-browser-warning": "true",
       ...(config?.headers || {}),
     },
-  }).then(async (data) => ({
-    ...data,
-    data: await data.text().then((val) => {
-      try {
-        return JSON.parse(val);
-      } catch (_) {
-        return val;
-      }
-    }),
-  }));
+  }).then(async (data) => {
+    if (mutate) {
+      return {
+        ...data,
+        resp: data,
+        data: await data.text().then((val) => {
+          try {
+            return JSON.parse(val);
+          } catch (_) {
+            return val;
+          }
+        }),
+      };
+    } else {
+      return { ...data, resp: data, data: "" };
+    }
+  });
 }
